// src/app/services/interfaces/dep.local.api.interface.ts
// Generated by https://quicktype.io

export interface LocalDep {
    id:           number;
    codigo:       null;
    nombre:       string;
    siglas:       string;
    naturaleza:   Naturaleza;
    idPadre:      number;
    esAMR:        null;
    amrId:        null;
    amrAlternaId: null;
    sector_CatId: null;
    activo:       boolean;
}

export enum Naturaleza {
    Empty = "-",
    Null = "null",
}